refactor: change directory structure
