+++ src/lib/StyledComponentsRegistry.tsx
